logs-service/requirements.txt
cassandra-driver==3.29.2
click==8.1.8
geomet==0.2.1.post1
kafka-python==2.0.2
six==1.17.0
